Add UnionValidationError
Add a rich (and correct, fixing the bugs in the existing one)
representation of errors raised when trying to validate an Union.

We store each individual exception in the causes attribute of the new
exception. This is not ideal because we lose the tracebacks.In the
future, that might be some sort of MultiError.