Enable ForwardRefs in dataclasses
In Python 3.14, type annotations are resolved in a delayed manner by
default, and get resolved, on demand, when accession __annotations__.
This allows writing recursive types such as a dataclass where some field
contains itself.

As a minor annoyance, accessing the type attribute of the fields, does
not resolve the delayed forward reference.

Add tests for Python 3.14